pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workflowbench"
version = "0.1.0"
description = "Lightweight benchmark harness for AI-driven business workflows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ajay", email = "hello@workflowbench.dev"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "jinja2>=3.1",
    "openai>=1.0",
    "anthropic>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[project.scripts]
workflowbench = "workflowbench.cli:main"

[project.urls]
Homepage = "https://github.com/thegeekajay/WorkflowBench"
Repository = "https://github.com/thegeekajay/WorkflowBench"
Issues = "https://github.com/thegeekajay/WorkflowBench/issues"

[tool.setuptools.packages.find]
include = ["workflowbench*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
